Changed addTodo, changeTodo, and added function toggleCompleted
addTodo now pushes an object into the todos array rather than a string, changeTodo reflects the changes made to addTodo and now takes a todoTextValue as an additional argument. Also added a new property and function called toggleCompleted which takes an argument of position which refers to an element in the todos array, and changes that element's completed status to the opposite of what it currently is.