Changed functionality of displayTodos,
Changed the way the displayTodos method outputs data by adding an 'if else' statement at the beginning of the method to check if the todos array's length is equal to '0', and if so to print a message to the console, else to run a loop which iterates over each item in the array and checks whether or not each element has a completed status of true, in which case it prints a '(x)' symbol next to the element. If the completed status is equal to false, it will print a symbol '( )' next to the element.