Added a 'toggleAll' function
Created a 'toggleAll' function that uses a 'for' loop to loop through all the elements of the 'todos' array, and check whether or not each one has a 'completed' status of true, and if it does increments the 'completedTodos' variable by 1.  Then it compares the value stored in variable 'completedTodos' to variable 'totalTodos' which is always equal to the total number of elements in the 'todos' array, and either changes all elements statuses to false if they're true, or vice-versa.